Crate spacetimedb_primitives

Modules

errno

Error numbers for the wasm abi.

Macros

col_list

Constructs a ColList like so col_list![0, 2].

errnos

Takes a macro that expects $($err_name:ident($errno:literal, $errmsg:literal),)* and invokes it with the errnos defined in this module.

Structs

ColId

The position of a column within a table.

ColSet

A compressed set of columns. Like a ColList, but guaranteed to be sorted and to contain no duplicate entries. Dereferences to a ColList for convenience.

ColumnAttribute

ConstraintId

An identifier for a constraint, unique within a database.

Constraints

Represents constraints for a database table. May apply to multiple columns.

IndexId

An identifier for an index, unique within a database.

ReducerId

The index of a reducer as defined in a module’s reducers list.

ScheduleId

An identifier for a schedule, unique within a database.

SequenceId

An identifier for a sequence, unique within a database.

TableId

An identifier for a table, unique within a database.

Enums

AttributeKind

The assigned constraint OR auto-inc marker for a Column

ColOrCols

A borrowed list of columns or a single one.

ConstraintKind

The assigned constraint for a Table

Constants

ROW_ITER_CHUNK_SIZE

The minimum size of a chunk yielded by a wasm abi RowIter.

Unions

ColList

This represents a list of ColIds but packed into a u64 in a way that takes advantage of the fact that in almost all cases, we won’t store a ColId larger than 62. In the rare case that we store larger ids, we fall back to a thin vec approach.