spacetimedb_primitives

Module errno

Error numbers for the wasm abi.

Constants

BSATN_DECODE_ERROR

Couldn’t decode the BSATN to the expected type

BUFFER_TOO_SMALL

The provided buffer is not large enough to store the data

HOST_CALL_FAILURE

ABI called by host returned an error

INDEX_NOT_UNIQUE

The index was not unique

NOT_IN_TRANSACTION

ABI call can only be made while in a transaction

NO_SPACE

The provided sink has no more space left

NO_SUCH_BYTES

The provided bytes source or sink is not valid

NO_SUCH_CONSOLE_TIMER

The provided console timer does not exist

NO_SUCH_INDEX

No such index

NO_SUCH_ITER

The provided row iterator is not valid

NO_SUCH_ROW

The row was not found, e.g., in an update call

NO_SUCH_TABLE

No such table

SCHEDULE_AT_DELAY_TOO_LONG

Specified delay in scheduling row was too long

UNIQUE_ALREADY_EXISTS

Value with given unique identifier already exists

Functions

strerror

Get the error message for an error number, if it exists.